+++ app/controllers/project_favorites_controller.rb
@@ -4,17 +4,26 @@
 class ProjectFavoritesController < ApplicationController
   before_action :authenticate_user!
   before_action :find_viewable_project_or_redirect
+  before_action :find_or_create_project_favorite
 
   # POST /project_favorites/1/colorpicker
   def colorpicker
-    project_favorite = @project.project_favorites.where(user_id: current_user.id).first_or_create
-    project_favorite.update color: params[:color]
+    @project_favorite.update project_favorite_params
     head :ok
   end
 
-  # POST /project_favorites/1/colorpicker
-  def favorite
-    project_favorite = @project.project_favorites.where(user_id: current_user.id).first_or_create
-    project_favorite.update favorite: (params[:favorite] == '1')
+  # PATCH /project_favorites/update?project_id=1
+  def update
+    @project_favorite.update project_favorite_params
+  end
+
+  private
+
+  def project_favorite_params
+    params.permit(:color, :favorite, :emails_enabled)
+  end
+
+  def find_or_create_project_favorite
+    @project_favorite = @project.project_favorites.where(user_id: current_user.id).first_or_create
   end
 end