+++ db/migrate/20160722141842_add_emails_enabled_to_users.rb
@@ -1,0 +1,6 @@
+class AddEmailsEnabledToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :emails_enabled, :boolean, null: false, default: false
+    add_index :users, :emails_enabled
+  end
+end