+++ db/migrate/20160722174804_add_emails_enabled_to_project_favorites.rb
@@ -1,0 +1,6 @@
+class AddEmailsEnabledToProjectFavorites < ActiveRecord::Migration
+  def change
+    add_column :project_favorites, :emails_enabled, :boolean, null: false, default: true
+    add_index :project_favorites, :emails_enabled
+  end
+end