+++ test/controllers/project_favorites_controller_test.rb
@@ -25,15 +25,16 @@ class ProjectFavoritesControllerTest < ActionController::TestCase
 
   test 'should create project favorite' do
     assert_difference('ProjectFavorite.where(favorite: true).count') do
-      post :favorite, project_id: projects(:one), favorite: '1', format: 'js'
+      patch :update, project_id: projects(:one), favorite: '1', format: 'js'
     end
     assert_not_nil assigns(:project)
-    assert_template 'favorite'
+    assert_template 'update'
+    assert_response :success
   end
 
   test 'should not create project favorite without valid id' do
     assert_difference('ProjectFavorite.where(favorite: true).count', 0) do
-      post :favorite, project_id: -1, favorite: '1', format: 'js'
+      patch :update, project_id: -1, favorite: '1', format: 'js'
     end
     assert_nil assigns(:project)
     assert_response :success
@@ -41,9 +42,30 @@ class ProjectFavoritesControllerTest < ActionController::TestCase
 
   test 'should remove project favorite' do
     assert_difference('ProjectFavorite.where(favorite: false).count') do
-      post :favorite, project_id: projects(:two), favorite: '0', format: 'js'
+      patch :update, project_id: projects(:two), favorite: '0', format: 'js'
     end
     assert_not_nil assigns(:project)
-    assert_template 'favorite'
+    assert_template 'update'
+    assert_response :success
+  end
+
+  test 'should enable project emails' do
+    login(users(:associated))
+    assert_difference('ProjectFavorite.where(emails_enabled: true).count') do
+      patch :update, project_id: projects(:one), emails_enabled: '1', format: 'js'
+    end
+    assert_not_nil assigns(:project)
+    assert_not_nil assigns(:project_favorite)
+    assert_template 'update'
+    assert_response :success
+  end
+
+  test 'should disable project emails' do
+    assert_difference('ProjectFavorite.where(emails_enabled: false).count') do
+      patch :update, project_id: projects(:one), emails_enabled: '0', format: 'js'
+    end
+    assert_not_nil assigns(:project)
+    assert_template 'update'
+    assert_response :success
   end
 end