+++ test/models/project_test.rb
@@ -2,21 +2,12 @@
 
 require 'test_helper'
 
+# Test project methods.
 class ProjectTest < ActiveSupport::TestCase
-  test 'should only send task creation email to users with email activated' do
-    assert_equal true, projects(:one).users_to_email(:sticky_creation).include?(users(:valid))
-    assert_equal false, projects(:one).users_to_email(:sticky_creation).include?(users(:send_no_email))
-    assert_equal false, projects(:one).users_to_email(:sticky_creation).include?(users(:send_no_email_for_project_one))
-    assert_equal false, projects(:one).users_to_email(:sticky_creation).include?(users(:send_no_email_for_sticky_creation))
-    assert_equal false, projects(:one).users_to_email(:sticky_creation).include?(users(:send_no_email_for_project_one_sticky_creation))
-  end
-
-  test 'should only send task completion email to users with email activated' do
-    assert_equal true, projects(:one).users_to_email(:sticky_completion).include?(users(:valid))
-    assert_equal false, projects(:one).users_to_email(:sticky_completion).include?(users(:send_no_email))
-    assert_equal false, projects(:one).users_to_email(:sticky_completion).include?(users(:send_no_email_for_project_one))
-    assert_equal false, projects(:one).users_to_email(:sticky_completion).include?(users(:send_no_email_for_sticky_completion))
-    assert_equal false, projects(:one).users_to_email(:sticky_completion).include?(users(:send_no_email_for_project_one_sticky_completion))
+  test 'should only send emails to users with email activated' do
+    assert_equal true, projects(:one).users_to_email.include?(users(:valid))
+    assert_equal false, projects(:one).users_to_email.include?(users(:send_no_email))
+    assert_equal false, projects(:one).users_to_email.include?(users(:send_no_email_for_project_one))
   end
 
   test 'viewer on project should see project' do